fix lint errors and update test
