+++ imports/blocks/give/fieldsets/confirm/SavedPaymentLayout.js
@@ -29,7 +29,7 @@ export default ({
   payment,
   header,
   children,
-  goToStepOne
+  goToStepOne,
 }: ISavedPaymentLayout) => {
   if (!billing || !payment) return null;
   const paymentInfo = {
@@ -56,7 +56,7 @@ export default ({
         <h6>
           <AccountNumber accountNumber={payment.cardNumber || payment.accountNumber} />
           &nbsp;&nbsp;
-          <Icon width="19px" height="12px" cardType={ cardType(paymentInfo, null) }/>
+          <Icon width="19px" height="12px" cardType={cardType(paymentInfo, null)} />
         </h6>
         <h6>{payment.expiration || payment.routingNumber}</h6>
         <h6>{payment.ccv}</h6>
