+++ imports/components/cards/__tests__/cards.SavedPayment.js
@@ -29,7 +29,7 @@ describe("SavedPaymentCard", () => {
     const mockOnClick = jest.fn();
     const tree = shallow(generateComponent({ payment: defaultAccount, onClick: mockOnClick}));
 
-    tree.simulate("click");
+    tree.find("[data-spec=\"saved-payment-close\"]").simulate("click");
     expect(mockOnClick).toHaveBeenCalledTimes(1);
   });
