+++ imports/store/give/saga/chargeTransaction.js
@@ -90,12 +90,12 @@ export default function* chargeTransaction({ state }) {
             variables: { token, name, id },
             updateQueries: {
               GivingDashboard: (prev, { mutationResult }) => {
-                const { savedPayment, error } = mutationResult.data.response;
-                if (error || !savedPayment) return prev;
+                const { savedPayment, success } = mutationResult.data.response;
+                if (!success || !savedPayment) return prev;
                 prev.savedPayments.push(savedPayment);
                 return prev;
-              }
-            }
+              },
+            },
           });
         } else {
           // XXX update data if returned
